+++ shotmanager/api/take.py
@@ -16,37 +16,36 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 """
-To do: module description here.
+Shot Manager API - Interface with take class
 """
 
-def get_name(take_instance):
+import bpy
+from shotmanager.properties.take import UAS_ShotManager_Take
+
+
+def get_name(take_instance: UAS_ShotManager_Take) -> str:
     return take_instance.name
 
 
-def set_name(take_instance, name):
-    """ Set a unique name to the take
-    """
+def set_name(take_instance: UAS_ShotManager_Take, name: str):
+    """Set a unique name to the take"""
     take_instance.name = name
 
 
-def get_name_path_compliant(take_instance):
+def get_name_path_compliant(take_instance: UAS_ShotManager_Take) -> str:
     return take_instance.getName_PathCompliant()
 
 
-def get_shot_list(take_instance, ignore_disabled=False):
-    """ Return a filtered copy of the shots associated to this take
-    """
+def get_shot_list(take_instance: UAS_ShotManager_Take, ignore_disabled: bool = False) -> list:
+    """Return a filtered copy of the shots associated to this take"""
     return take_instance.getShotList(ignoreDisabled=ignore_disabled)
 
 
-def get_num_shots(take_instance, ignore_disabled=False):
-    """ Return the number of shots of the take
-    """
+def get_num_shots(take_instance: UAS_ShotManager_Take, ignore_disabled: bool = False) -> int:
+    """Return the number of shots of the take"""
     return take_instance.getNumShots(ignoreDisabled=ignore_disabled)
 
 
-def get_shots_using_camera(take_instance, cam, ignore_disabled=False):
-    """ Return the list of all the shots used by the specified camera
-    """
+def get_shots_using_camera(take_instance: UAS_ShotManager_Take, cam: bpy.types.Camera, ignore_disabled: bool = False):
+    """Return the list of all the shots used by the specified camera"""
     return take_instance.getShotsUsingCamera(cam, ignoreDisabled=ignore_disabled)
-